requirements.txt
python-telegram-bot==13.15
python-dotenv
gspread
oauth2client
matplotlib
yfinance
feedparser
forex-python
google-generativeai
